+++ ricaun.Revit.UI/RibbonItemExtension.cs
@@ -149,7 +149,7 @@ namespace ricaun.Revit.UI
 
         #region Set RibbonButton
         /// <summary>
-        /// Set RibbonButton/ComboBox/TextBox Image
+        /// Set RibbonButton/ComboBox/ComboBoxMember/TextBox Image
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -166,13 +166,13 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton/ComboBox/TextBox LargeImage
+        /// Set RibbonButton/ComboBox/ComboBoxMember/TextBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
         /// <param name="largeImage"></param>
         /// <returns></returns>
-        /// <remarks>When <see cref="ComboBox"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
+        /// <remarks>When <see cref="ComboBox"/>, <see cref="ComboBoxMember"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, string largeImage) where TRibbonItem : RibbonItem
         {
             var bitmapSource = largeImage?.GetBitmapSource();
@@ -184,7 +184,7 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton/ComboBox/TextBox Image
+        /// Set RibbonButton/ComboBox/ComboBoxMember/TextBox Image
         /// </summary>
         /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
         /// <param name="ribbonItem"></param>
@@ -198,6 +198,9 @@ namespace ricaun.Revit.UI
             else if (ribbonItem is ComboBox comboBox)
                 comboBox.Image = image?.GetBitmapFrame(16, (frame) => { comboBox.Image = frame; });
 
+            else if (ribbonItem is ComboBoxMember comboBoxMember)
+                comboBoxMember.Image = image?.GetBitmapFrame(16, (frame) => { comboBoxMember.Image = frame; });
+
             else if (ribbonItem is TextBox textBox)
                 textBox.Image = image?.GetBitmapFrame(16, (frame) => { textBox.Image = frame; });
 
@@ -205,13 +208,13 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton/ComboBox/TextBox LargeImage
+        /// Set RibbonButton/ComboBox/ComboBoxMember/TextBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
         /// <param name="ribbonItem"></param>
         /// <param name="largeImage"></param>
         /// <returns></returns>
-        /// <remarks>When <see cref="ComboBox"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
+        /// <remarks>When <see cref="ComboBox"/>, <see cref="ComboBoxMember"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItem
         {
             if (ribbonItem is RibbonButton ribbonButton)
@@ -223,6 +226,9 @@ namespace ricaun.Revit.UI
 
             else if (ribbonItem is ComboBox comboBox)
                 comboBox.SetImage(largeImage);
+
+            else if (ribbonItem is ComboBoxMember comboBoxMember)
+                comboBoxMember.SetImage(largeImage);
 
             else if (ribbonItem is TextBox textBox)
                 textBox.SetImage(largeImage);