+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change `GetRibbonItem` to `GetRibbonItem_Alternative` to fix null when panel is removed.
 - Update `SetImage` to work with `ComboBoxMember`
 - Add `CreateComboBoxMember` to create `ComboBoxMember`.
-- Add `RibbonThemeBitmapUtils` to change theme for `BitmapSource`.
+- Add `RibbonThemeImageUtils` to change theme for `ImageSource`.
+- Update `RibbonThemeImageUtils` with public `GetThemeImageSource`.
 - Add `RibbonThemePanelUtils` to update the theme for itens in the `RibbonPanel`.
 - Update `RibbonPanel` create and remove to update the theme of the itens.
 ### Tests